requirements.txt
absl-py==2.1.0
addict==2.4.0
alembic==1.13.3
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
basicsr==1.4.2
beautifulsoup4==4.12.2
binarytree==6.5.1
bs4==0.0.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
decouple==0.0.7
defusedxml==0.8.0rc2
dj-database-url==2.3.0
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.1.0
django-cors-headers==4.5.0
django-extensions==3.2.3
django-filter==24.3
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-yasg==1.21.8
executing==2.1.0
facexlib==0.3.0
filelock==3.16.1
filterpy==1.4.5
fonttools==4.54.1
fpdf==1.7.2
fsspec==2024.10.0
future==1.0.0
gfpgan==1.3.8
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
lightgbm==4.5.0
llvmlite==0.43.0
lmdb==1.5.1
lxml==5.3.0
Mako==1.3.6
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
meteostat==1.6.8
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
opencv-python==4.10.0.84
optuna==4.0.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
psycopg==3.2.3
psycopg2==2.9.10
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.10.1
pyngrok==7.2.3
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
realesrgan==0.3.0
realesrgan-ncnn-py==2.0.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.27.1
sendgrid==3.6.5
sendgrid-django==4.2.0
setuptools==75.3.0
setuptools-scm==8.1.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.1
tb-nightly==2.19.0a20241106
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli==2.0.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.2
whitenoise==6.8.1
widgetsnbextension==4.0.13
wsproto==1.2.0
xgboost==2.1.2
yapf==0.40.2
zipp==3.20.2
